refactor: infix integration tests
